Deprecate XPLINKCallDescriptorSnippet
- Migrate `generateCallDescriptorValue` to linkage class
- Stop compiling `XPLINKCallDescriptorSnippet.cpp` but leave the
compilation unit intact until downstream projects remove it from their
makefiles
