+++ compiler/z/codegen/SystemLinkagezOS.cpp
@@ -71,7 +71,6 @@
 #include "z/codegen/S390GenerateInstructions.hpp"
 #include "z/codegen/S390Instruction.hpp"
 #include "z/codegen/SystemLinkagezOS.hpp"
-#include "z/codegen/snippet/XPLINKCallDescriptorSnippet.hpp"
 #include "OMR/Bytes.hpp"
 
 #ifdef J9_PROJECT_SPECIFIC
@@ -649,7 +648,7 @@ TR::S390zOSSystemLinkage::genCallNOPAndDescriptor(TR::Instruction* cursor, TR::N
       cursor = generateS390LabelInstruction(cg(), InstOpCode::LABEL, node, xplinkCallDescriptorBeginLabel, cursor);
       cursor = generateDataConstantInstruction(cg(), TR::InstOpCode::DC, node, 0x00000000, cursor);
 
-      uint32_t callDescriptorValue = TR::XPLINKCallDescriptorSnippet::generateCallDescriptorValue(this, callNode);
+      uint32_t callDescriptorValue = generateCallDescriptorValue(callNode);
       cursor = generateDataConstantInstruction(cg(), TR::InstOpCode::DC, node, callDescriptorValue, cursor);
       cursor = generateS390LabelInstruction(cg(), InstOpCode::LABEL, node, xplinkCallDescriptorEndLabel, cursor);
       }
@@ -662,6 +661,177 @@ TR::S390zOSSystemLinkage::genCallNOPAndDescriptor(TR::Instruction* cursor, TR::N
       }
 
    return cursor;
+   }
+
+uint32_t
+TR::S390zOSSystemLinkage::generateCallDescriptorValue(TR::Node* callNode)
+   {
+   uint32_t result = 0;
+
+   if (cg()->comp()->target().is32Bit())
+      {
+      uint32_t returnValueAdjust = 0;
+
+      // 5 bit values for Return Value Adjust field of XPLLINK descriptor
+      enum ReturnValueAdjust
+         {
+         XPLINK_RVA_RETURN_VOID_OR_UNUSED  = 0x00,
+         XPLINK_RVA_RETURN_INT32_OR_LESS   = 0x01,
+         XPLINK_RVA_RETURN_INT64           = 0x02,
+         XPLINK_RVA_RETURN_FAR_POINTER     = 0x04,
+         XPLINK_RVA_RETURN_FLOAT4          = 0x08,
+         XPLINK_RVA_RETURN_FLOAT8          = 0x09,
+         XPLINK_RVA_RETURN_FLOAT16         = 0x0A,
+         XPLINK_RVA_RETURN_COMPLEX4        = 0x0C,
+         XPLINK_RVA_RETURN_COMPLEX8        = 0x0D,
+         XPLINK_RVA_RETURN_COMPLEX16       = 0x0E,
+         XPLINK_RVA_RETURN_AGGREGATE       = 0x10,
+         };
+
+      TR::DataType dataType = callNode->getDataType();
+
+      switch (dataType)
+         {
+         case TR::NoType:
+            returnValueAdjust = XPLINK_RVA_RETURN_VOID_OR_UNUSED;
+            break;
+
+         case TR::Int8:
+         case TR::Int16:
+         case TR::Int32:
+         case TR::Address:
+            returnValueAdjust = XPLINK_RVA_RETURN_INT32_OR_LESS;
+            break;
+
+         case TR::Int64:
+            returnValueAdjust = XPLINK_RVA_RETURN_INT64;
+            break;
+
+         case TR::Float:
+            returnValueAdjust = XPLINK_RVA_RETURN_FLOAT4;
+            break;
+
+         case TR::Double:
+            returnValueAdjust = XPLINK_RVA_RETURN_FLOAT8;
+            break;
+
+         default:
+            TR_ASSERT_FATAL(false, "Unknown datatype (%s) for call node (%p)", dataType.toString(), callNode);
+            break;
+         }
+
+      result |= returnValueAdjust << 24;
+
+      //
+      // Float parameter description fields
+      // Bits 8-31 inclusive
+      //
+
+      uint32_t parmAreaOffset = 0;
+
+#ifdef J9_PROJECT_SPECIFIC
+      TR::MethodSymbol* callSymbol = callNode->getSymbol()->castToMethodSymbol();
+      if (callSymbol->isJNI() && callNode->isPreparedForDirectJNI())
+         {
+         TR::ResolvedMethodSymbol * cs = callSymbol->castToResolvedMethodSymbol();
+         TR_ResolvedMethod * resolvedMethod = cs->getResolvedMethod();
+         // JNI Calls include a JNIEnv* pointer that is not included in list of children nodes.
+         // For FastJNI, certain calls do not require us to pass the JNIEnv.
+         if (!cg()->fej9()->jniDoNotPassThread(resolvedMethod))
+            parmAreaOffset += sizeof(uintptrj_t);
+
+         // For FastJNI, certain calls do not have to pass in receiver object.
+         if (cg()->fej9()->jniDoNotPassReceiver(resolvedMethod))
+            parmAreaOffset -= sizeof(uintptrj_t);
+         }
+#endif
+
+      uint32_t parmDescriptorFields = 0;
+
+      TR::Symbol *funcSymbol = callNode->getSymbolReference()->getSymbol();
+
+      uint32_t firstArgumentChild = callNode->getFirstArgumentIndex();
+      int32_t to = callNode->getNumChildren() - 1;
+      int32_t parmCount = 1;
+
+      int32_t floatParmNum = 0;
+      uint32_t gprSize = cg()->machine()->getGPRSize();
+
+      uint32_t lastFloatParmAreaOffset = 0;
+
+      bool done = false;
+      for (int32_t i = firstArgumentChild; (i <= to) && !done; i++, parmCount++)
+         {
+         TR::Node *child = callNode->getChild(i);
+         TR::DataType dataType = child->getDataType();
+         TR::SymbolReference *parmSymRef = child->getOpCode().hasSymbolReference() ? child->getSymbolReference() : NULL;
+         int32_t argSize = 0;
+
+         if (parmSymRef == NULL)
+            argSize = child->getSize();
+         else
+            argSize = parmSymRef->getSymbol()->getSize();
+
+
+         // Note: complex type is attempted to be handled although other code needs
+         // to change in 390 codegen to support complex
+         //
+         // PERFORMANCE TODO: it is desirable to use the defined "parameter count" of
+         // the function symbol to help determine if we have an unprototyped argument
+         // of a call (site) to a vararg function.  Currently we overcompensate for
+         // outgoing float parms to vararg functions and always shadow in FPR and
+         // and stack/gprs as with an unprotoyped call - see pushArg(). Precise
+         // information can help remove such compensation. Changes to fix this would
+         // involve: this function, pushArg() and buildArgs().
+
+         int32_t numFPRsNeeded = 0;
+         switch (dataType)
+            {
+            case TR::Float:
+            case TR::Double:
+#ifdef J9_PROJECT_SPECIFIC
+            case TR::DecimalFloat:
+            case TR::DecimalDouble:
+#endif
+               numFPRsNeeded = 1;
+               break;
+#ifdef J9_PROJECT_SPECIFIC
+            case TR::DecimalLongDouble:
+               break;
+#endif
+            }
+
+         if (numFPRsNeeded != 0)
+            {
+            uint32_t unitSize = argSize / numFPRsNeeded;
+            uint32_t wordsToPreviousParm = (parmAreaOffset - lastFloatParmAreaOffset) / gprSize;
+            if (wordsToPreviousParm > 0xF)
+               { // to big for descriptor. Will pass in stack
+               done = true; // done
+               }
+            uint32_t val = wordsToPreviousParm + ((unitSize == 4) ? 0x10 : 0x20);
+
+            parmDescriptorFields |= val << (6 * (3 - floatParmNum));
+
+            floatParmNum++;
+
+            if (floatParmNum >= self()->getNumFloatArgumentRegisters())
+               {
+               done = true;
+               }
+            }
+         parmAreaOffset += argSize < gprSize ? gprSize : argSize;
+
+         if (numFPRsNeeded != 0)
+            {
+            lastFloatParmAreaOffset = parmAreaOffset;
+            }
+         }
+
+      result |= parmDescriptorFields;
+      }
+
+   return result;
    }
 
 TR::Instruction *
