+++ compiler/z/codegen/snippet/XPLINKCallDescriptorSnippet.hpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019, 2019 IBM Corp. and others
+ * Copyright (c) 2019, 2020 IBM Corp. and others
  *
  * This program and the accompanying materials are made available under
  * the terms of the Eclipse Public License 2.0 which accompanies this
@@ -62,8 +62,6 @@ class XPLINKCallDescriptorSnippet : public TR::S390ConstantDataSnippet
     *     Size (in bytes) of the call descriptor data structure.
     */
    static const size_t SIZE = 8;
-
-   static uint32_t generateCallDescriptorValue(TR::S390zOSSystemLinkage* linkage, TR::Node* callNode);
 
    public:
